Adding sink service for logging acm/idm requests
The ACM/IDM flow goes through https.  This means the http logger can't
automatically log this traffic in an inspectable way.  We forward the
requests and the responses to the sink service so we have a good idea
of what is going on in the standard setup.  This approach should make
it easier to monitor misuse on production systems.