Further cleanup to development setup
The development setup does not need default ports to be open for the
majority of the microservices.  Services also don't need to run in
development mode in the majority of cases as it will most likely be
configured when the respective volume is mounted and the ports are
opened up.