fix: refine query to match MR cases
We need to match both those cases that don't have a publication-flow,
i.e. cases created via the "Dossiers" tab that haven't been published
yet via an agenda. For the cases that do have a publication-flow, we
only want to match those that are related to an agenda. Before we only
matched the latter, which means we missed out on a ton of cases.