chore: add solitary public as eager index group
Not strictly necessary, but in (the very rare!) case where a user is
active on the app and gets logged out (may happen on ACC after a data
sync), a search query will be sent with allowed groups `public` and
`clean`, which without this extra group won't match and will trigger an
indexation (which would finish in seconds, but still).

This is mostly for completeness and isn't /really/ needed.